requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
scikit-learn==1.3.2
pandas==2.1.3
joblib==1.3.2
numpy==1.26.2
matplotlib==3.8.2
imbalanced-learn==0.11.0
python-dotenv==1.0.0
slowapi==0.1.9
PyMySQL==1.1.0
shap==0.43.0
passlib==1.7.4
bcrypt==4.1.1
PyJWT==2.8.1
sqlalchemy==2.0.23
cachetools==5.3.1
